refactor(admin): extract kiosk setup form into reusable KioskForm component
