+++ src/views/admin/KioskManagement.tsx
@@ -1,4 +1,6 @@
-import React, { useState, useEffect, useRef } from 'react';
+"use client";
+
+import React, { useState, useEffect } from 'react';
 import { db } from '../../shared/lib/firebase';
 import { useKiosks } from '../../shared/lib/hooks/useKiosks';
 import { useCampaigns } from '../../entities/campaign';
@@ -17,17 +19,16 @@ import { Screen, Kiosk, AdminSession, Permission } from '../../shared/types';
 // UI Components
 import { Button } from '../../shared/ui/button';
 import { Input } from '../../shared/ui/input';
-import { Label } from '../../shared/ui/label';
 import { Badge } from '../../shared/ui/badge';
 import { Card, CardContent } from '../../shared/ui/card';
 import { Dialog, DialogContent, DialogTitle, VisuallyHidden } from '../../shared/ui/dialog';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../shared/ui/table';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../../shared/ui/select';
 
 import {
-  Plus, Edit, Trash2, MapPin, AlertTriangle, Save, Eye, EyeOff, Copy, Check, ArrowLeft, Search, DollarSign, Users, Menu, X, Grid3X3, List, GalleryThumbnails, Image, Settings, Activity
+  Plus, Edit, Trash2, MapPin, Eye, EyeOff, Copy, Check, ArrowLeft, Search, DollarSign, Users, Settings, Activity, AlertTriangle
 } from 'lucide-react';
 import { AdminLayout } from './AdminLayout';
+import { KioskForm, KioskFormData } from './components/KioskForm';
 
 
 export function KioskManagement({ onNavigate, onLogout, userSession, hasPermission }: {
@@ -73,7 +74,7 @@ export function KioskManagement({ onNavigate, onLogout, userSession, hasPermissi
 
   const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
   const [editingKiosk, setEditingKiosk] = useState<Kiosk | null>(null);
-  const [newKiosk, setNewKiosk] = useState({ 
+  const [newKiosk, setNewKiosk] = useState<KioskFormData>({ 
     name: '', 
     location: '', 
     accessCode: '', 
@@ -81,25 +82,6 @@ export function KioskManagement({ onNavigate, onLogout, userSession, hasPermissi
     assignedCampaigns: [] as string[],
     displayLayout: 'grid' as 'grid' | 'list' | 'carousel'
   });
-  
-  // Dialog state for mobile navigation
-  const [activeSection, setActiveSection] = useState('basic-info');
-  const [isMobileSidebarOpen, setIsMobileSidebarOpen] = useState(false);
-  
-  // Refs for scroll container and sections
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-  const sectionRefs = {
-    'basic-info': useRef<HTMLElement>(null),
-    'campaigns': useRef<HTMLElement>(null),
-    'display': useRef<HTMLElement>(null)
-  };
-  
-  // Navigation items for dialog
-  const navigationItems = [
-    { id: 'basic-info', label: 'BASIC INFO' },
-    { id: 'campaigns', label: 'CAMPAIGNS' },
-    { id: 'display', label: 'DISPLAY' }
-  ];
   
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
   const [kioskToDelete, setKioskToDelete] = useState<Kiosk | null>(null);
@@ -112,102 +94,6 @@ export function KioskManagement({ onNavigate, onLogout, userSession, hasPermissi
     refreshKiosks();
     refreshCampaigns();
   }, [refreshKiosks, refreshCampaigns]);
-
-  // IntersectionObserver for scroll-synced sidebar highlighting
-  useEffect(() => {
-    if (!isCreateDialogOpen) return;
-
-    // Use a timeout to ensure DOM is ready
-    const timeoutId = setTimeout(() => {
-      const container = scrollContainerRef.current;
-      if (!container) return;
-
-      const observer = new IntersectionObserver(
-        (entries) => {
-          // Find the entry with the highest intersection ratio
-          let maxRatio = 0;
-          let activeId = '';
-          
-          entries.forEach((entry) => {
-            if (entry.intersectionRatio > maxRatio) {
-              maxRatio = entry.intersectionRatio;
-              activeId = entry.target.id;
-            }
-          });
-          
-          // Only update if we have a meaningful intersection
-          if (maxRatio > 0.1 && activeId) {
-            setActiveSection(activeId);
-          }
-        },
-        {
-          root: container,
-          rootMargin: '0px 0px -50% 0px',
-          threshold: [0, 0.1, 0.25, 0.5, 0.75, 1.0]
-        }
-      );
-
-      // Observe all sections
-      Object.values(sectionRefs).forEach((ref) => {
-        if (ref.current) {
-          observer.observe(ref.current);
-        }
-      });
-
-      // Cleanup function
-      const cleanup = () => {
-        observer.disconnect();
-      };
-
-      return cleanup;
-    }, 200);
-
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [isCreateDialogOpen]);
-
-  // Debug: Check refs when dialog state changes
-  useEffect(() => {
-    if (isCreateDialogOpen) {
-      console.log('🚪 Dialog opened - checking refs after render');
-      setTimeout(() => {
-        console.log('⏰ Delayed ref check:');
-        console.log('📦 Scroll container:', scrollContainerRef.current);
-        Object.entries(sectionRefs).forEach(([id, ref]) => {
-          console.log(`📍 Section ${id}:`, ref.current);
-          if (ref.current) {
-            const rect = ref.current.getBoundingClientRect();
-            console.log(`📏 Section ${id} dimensions:`, {
-              width: rect.width,
-              height: rect.height,
-              top: rect.top,
-              visible: rect.height > 0 && rect.width > 0
-            });
-          }
-        });
-      }, 100);
-    }
-  }, [isCreateDialogOpen]);
-
-  const scrollToSection = (sectionId: string) => {
-    const sectionRef = sectionRefs[sectionId as keyof typeof sectionRefs];
-    if (sectionRef?.current) {
-      sectionRef.current.scrollIntoView({ 
-        behavior: 'smooth', 
-        block: 'start' 
-      });
-      setActiveSection(sectionId);
-    }
-  };
-
-  // Handle keyboard navigation
-  const handleNavKeyDown = (event: React.KeyboardEvent, sectionId: string) => {
-    if (event.key === 'Enter' || event.key === ' ') {
-      event.preventDefault();
-      scrollToSection(sectionId);
-    }
-  };
 
   const handleAssignCampaign = (campaignId: string) => {
     if (needsOnboarding) {
@@ -279,11 +165,15 @@ export function KioskManagement({ onNavigate, onLogout, userSession, hasPermissi
       setNewKiosk({ name: '', location: '', accessCode: '', status: 'offline', assignedCampaigns: [], displayLayout: 'grid' });
       setIsCreateDialogOpen(false);
       setEditingKiosk(null);
-      setActiveSection('basic-info');
-      setIsMobileSidebarOpen(false);
     } catch (error) {
       console.error("Error saving kiosk: ", error);
     }
+  };
+
+  const handleCancel = () => {
+    setIsCreateDialogOpen(false);
+    setEditingKiosk(null);
+    setNewKiosk({ name: '', location: '', accessCode: '', status: 'offline', assignedCampaigns: [], displayLayout: 'grid' });
   };
 
   const handleEditKiosk = (kiosk: Kiosk) => {
@@ -620,560 +510,25 @@ export function KioskManagement({ onNavigate, onLogout, userSession, hasPermissi
       </div>
 
       {/* Kiosk Setup/Edit Dialog */}
-      <Dialog 
-        open={isCreateDialogOpen} 
+      <KioskForm
+        open={isCreateDialogOpen}
         onOpenChange={(open) => {
           setIsCreateDialogOpen(open);
           if (!open) {
-            // Reset state when dialog closes
-            setActiveSection('basic-info');
-            setIsMobileSidebarOpen(false);
             setEditingKiosk(null);
             setNewKiosk({ name: '', location: '', accessCode: '', status: 'offline', assignedCampaigns: [], displayLayout: 'grid' });
           }
         }}
-      >
-        <DialogContent className="sm:max-w-6xl p-0 border-0 shadow-2xl bg-white rounded-2xl overflow-hidden font-lexend h-[90vh] w-[95vw] sm:w-full flex flex-col [&>button]:hidden sm:[&>button]:flex">
-          <VisuallyHidden>
-            <DialogTitle>{editingKiosk ? 'Edit Kiosk Configuration' : 'Kiosk Setup Configuration'}</DialogTitle>
-          </VisuallyHidden>
-          
-          {/* Mobile Header - Fixed */}
-          <div className="sm:hidden flex flex-col border-b border-gray-200 bg-gradient-to-r from-gray-50 to-gray-100 flex-shrink-0">
-            {/* Header with aligned controls and title */}
-            <div className="flex items-center justify-between px-4 pt-3 pb-4">
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => setIsMobileSidebarOpen(!isMobileSidebarOpen)}
-                className="p-3 hover:bg-white/60 rounded-xl border border-gray-200/50 shadow-sm transition-all duration-200 hover:shadow-md"
-              >
-                <Menu className="w-5 h-5 text-gray-700" />
-              </Button>
-              
-              {/* Centered Title Section */}
-              <div className="text-center flex-1 mx-4">
-                <h3 className="text-lg font-semibold text-gray-900 mb-1">{editingKiosk ? 'Edit Kiosk' : 'Kiosk Setup'}</h3>
-                <p className="text-sm text-gray-600">Configuration</p>
-              </div>
-              
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => {
-                  setIsCreateDialogOpen(false);
-                  setActiveSection('basic-info');
-                  setIsMobileSidebarOpen(false);
-                }}
-                className="p-3 hover:bg-white/60 rounded-xl border border-gray-200/50 shadow-sm transition-all duration-200 hover:shadow-md"
-              >
-                <X className="w-5 h-5 text-gray-700" />
-              </Button>
-            </div>
-          </div>
-
-          {/* Main Content Area - Flex container */}
-          <div className="flex flex-col sm:flex-row flex-1 min-h-0">
-            {/* Sidebar Navigation */}
-            <nav 
-              className={`${
-                isMobileSidebarOpen ? 'block' : 'hidden'
-              } sm:block w-full sm:w-72 bg-gradient-to-b from-gray-50 to-gray-100 sm:bg-gray-50 border-r border-gray-200 p-6 sm:p-8 absolute sm:relative z-10 sm:z-auto h-full sm:h-auto flex-shrink-0 shadow-xl sm:shadow-none`} 
-              aria-label="Kiosk setup navigation"
-            >
-              {/* Mobile Navigation Header */}
-              <div className="sm:hidden mb-6 pb-4 border-b border-gray-200">
-                <div>
-                  <h3 className="text-lg font-semibold text-gray-900">Navigation</h3>
-                  <p className="text-sm text-gray-600 mt-1">Choose a section</p>
-                </div>
-              </div>
-
-              {/* Desktop Header */}
-              <div className="hidden sm:block mb-8">
-                <h3 className="text-lg font-semibold text-gray-900 mb-2">Kiosk</h3>
-                <p className="text-sm text-gray-500">Configuration</p>
-              </div>
-              
-              <div className="relative">
-                {/* Navigation Items */}
-                <ul className="space-y-3 relative" role="list">
-                  {navigationItems.map((item) => (
-                    <li key={item.id}>
-                      <button
-                        onClick={() => {
-                          scrollToSection(item.id);
-                          setIsMobileSidebarOpen(false);
-                        }}
-                        onKeyDown={(e) => handleNavKeyDown(e, item.id)}
-                        className={`w-full flex items-center gap-3 px-4 sm:px-5 py-4 rounded-xl cursor-pointer transition-all duration-200 text-left shadow-sm ${
-                          activeSection === item.id 
-                            ? 'bg-green-600 text-white shadow-lg shadow-green-600/25 border border-green-500' 
-                            : 'text-gray-700 hover:bg-white hover:shadow-md border border-gray-200/50 bg-gray-50/50'
-                        }`}
-                        aria-current={activeSection === item.id ? 'step' : undefined}
-                        tabIndex={0}
-                      >
-                        <div className="text-sm font-medium">{item.label}</div>
-                        {activeSection === item.id && (
-                          <div className="ml-auto w-2 h-2 bg-white rounded-full"></div>
-                        )}
-                      </button>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            </nav>
-
-            {/* Main Content Column - Flex container */}
-            <div className="flex-1 flex flex-col min-h-0 min-w-0">
-              {/* Header - Fixed */}
-              <header className="hidden sm:flex items-center justify-between p-6 lg:p-8 border-b border-gray-200 flex-shrink-0">
-                <div className="flex items-center gap-4">
-                  <div className="text-base text-gray-500 uppercase tracking-wide font-medium">
-                    {editingKiosk ? 'EDIT • KIOSK' : 'SETUP • NEW KIOSK'}
-                  </div>
-                </div>
-              </header>
-
-              {/* Scrollable Content Container - The only scroll area */}
-              <div 
-                ref={scrollContainerRef}
-                className="flex-1 overflow-y-auto min-h-0"
-              >
-                {/* Basic Information Section */}
-                <section 
-                  id="basic-info"
-                  ref={sectionRefs['basic-info']}
-                  className="p-4 sm:p-6 lg:p-8 border-b border-gray-100"
-                >
-                  <div className="max-w-4xl">
-                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-900 mb-6 sm:mb-8">
-                      Basic Information
-                    </h2>
-                    
-                    <div className="space-y-6">
-                      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                        <div className="lg:col-span-2">
-                          <Label htmlFor="kioskName" className="text-sm font-medium text-gray-700 mb-2 block">
-                            KIOSK NAME
-                          </Label>
-                          <Input
-                            id="kioskName"
-                            value={newKiosk.name}
-                            onChange={(e) => setNewKiosk(p => ({ ...p, name: e.target.value }))}
-                            placeholder="e.g. Reception Donation Terminal"
-                            className="h-12 text-base border-gray-300 rounded-lg focus:border-green-500 focus:ring-green-500"
-                          />
-                        </div>
-
-                        <div className="lg:col-span-2">
-                          <Label htmlFor="kioskLocation" className="text-sm font-medium text-gray-700 mb-2 block">
-                            PHYSICAL LOCATION
-                          </Label>
-                          <Input
-                            id="kioskLocation"
-                            value={newKiosk.location}
-                            onChange={(e) => setNewKiosk(p => ({ ...p, location: e.target.value }))}
-                            placeholder="e.g. Main Foyer, Ground Floor, Near Lifts"
-                            className="h-12 text-base border-gray-300 rounded-lg focus:border-green-500 focus:ring-green-500"
-                          />
-                        </div>
-
-                        <div>
-                          <Label htmlFor="accessCode" className="text-sm font-medium text-gray-700 mb-2 block">
-                            ACCESS CODE / PAIRING KEY
-                          </Label>
-                          <Input
-                            id="accessCode"
-                            value={newKiosk.accessCode}
-                            onChange={(e) => setNewKiosk(p => ({ ...p, accessCode: e.target.value }))}
-                            placeholder="e.g. DON-001-UK"
-                            className="h-12 text-base border-gray-300 rounded-lg focus:border-green-500 focus:ring-green-500"
-                          />
-                        </div>
-
-                        <div>
-                          <Label htmlFor="status" className="text-sm font-medium text-gray-700 mb-2 block">
-                            STATUS
-                          </Label>
-                          <Select
-                            value={newKiosk.status}
-                            onValueChange={(value: Kiosk['status']) => setNewKiosk(p => ({ ...p, status: value }))}
-                          >
-                            <SelectTrigger className="h-12 text-base border-gray-300 rounded-lg focus:border-green-500 focus:ring-green-500">
-                              <SelectValue placeholder="Select status" />
-                            </SelectTrigger>
-                            <SelectContent>
-                              <SelectItem value="online">Active</SelectItem>
-                              <SelectItem value="offline">Inactive</SelectItem>
-                              <SelectItem value="maintenance">Maintenance</SelectItem>
-                            </SelectContent>
-                          </Select>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </section>
-
-                {/* Campaigns Section */}
-                <section 
-                  id="campaigns"
-                  ref={sectionRefs['campaigns']}
-                  className="p-4 sm:p-6 lg:p-8 border-b border-gray-100"
-                >
-                  <div className="max-w-4xl">
-                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-900 mb-6 sm:mb-8">
-                      Campaigns
-                    </h2>
-                    
-                    {/* Assigned Campaigns Section */}
-                    <div className="mb-6 sm:mb-8">
-                      <div className="flex items-center gap-2 mb-4">
-                        <div className="w-6 h-6 rounded-full border-2 border-green-500 flex items-center justify-center">
-                          <div className="w-2 h-2 bg-green-500 rounded-full"></div>
-                        </div>
-                        <h3 className="text-base sm:text-lg font-medium text-gray-900">Assigned Campaigns</h3>
-                        <Badge variant="secondary" className="bg-gray-100 text-gray-600 text-xs">
-                          {newKiosk.assignedCampaigns.length}
-                        </Badge>
-                      </div>
-                      
-                      <div className="max-h-64 sm:max-h-80 overflow-y-auto">
-                        {newKiosk.assignedCampaigns.length === 0 ? (
-                          <div className="border-2 border-dashed border-gray-200 rounded-lg p-6 sm:p-8 text-center">
-                            <button
-                              onClick={() => {
-                                const availableCampaignsElement = document.getElementById('available-campaigns');
-                                if (availableCampaignsElement) {
-                                  availableCampaignsElement.scrollIntoView({ 
-                                    behavior: 'smooth', 
-                                    block: 'start' 
-                                  });
-                                }
-                              }}
-                              className="w-10 h-10 sm:w-12 sm:h-12 bg-gray-100 hover:bg-green-100 rounded-full flex items-center justify-center mx-auto mb-3 transition-colors cursor-pointer group"
-                              aria-label="Scroll to available campaigns"
-                            >
-                              <Plus className="w-5 h-5 sm:w-6 sm:h-6 text-gray-400 group-hover:text-green-600 transition-colors" />
-                            </button>
-                            <h4 className="text-base sm:text-lg font-medium text-gray-900 mb-2">No campaigns assigned</h4>
-                            <p className="text-gray-500 text-sm">Select campaigns from the available list below to get started</p>
-                          </div>
-                        ) : (
-                          <div className="space-y-3">
-                            {newKiosk.assignedCampaigns.map(campaignId => {
-                              const campaign = campaigns.find(c => c.id === campaignId);
-                              if (!campaign) return null;
-                              
-                              const fundingPercentage = Math.round((campaign.raised / campaign.goal) * 100);
-                              
-                              return (
-                                <div key={campaign.id} className="flex flex-col sm:flex-row sm:items-center gap-3 sm:gap-4 p-4 bg-white border border-gray-200 rounded-lg">
-                                  <div className="flex items-center gap-3 sm:gap-4 flex-1 min-w-0">
-                                    <div className="w-12 h-12 bg-gray-100 rounded-lg flex items-center justify-center flex-shrink-0">
-                                      {campaign.coverImageUrl ? (
-                                        <img 
-                                          src={campaign.coverImageUrl} 
-                                          alt={campaign.title}
-                                          className="w-full h-full object-cover rounded-lg"
-                                        />
-                                      ) : (
-                                        <span className="text-xs font-medium text-gray-500">
-                                          {campaign.title.substring(0, 3).toUpperCase()}
-                                        </span>
-                                      )}
-                                    </div>
-                                    
-                                    <div className="flex-1 min-w-0">
-                                      <h4 className="font-medium text-gray-900 truncate text-sm sm:text-base">{campaign.title}</h4>
-                                      <div className="flex flex-col sm:flex-row sm:items-center gap-1 sm:gap-4 text-xs sm:text-sm text-gray-500 mt-1">
-                                        <span>{formatCurrency(campaign.raised)} raised</span>
-                                        <span className="hidden sm:inline">•</span>
-                                        <span>{fundingPercentage}% funded</span>
-                                      </div>
-                                    </div>
-                                  </div>
-                                  
-                                  <Button
-                                    variant="outline"
-                                    size="sm"
-                                    onClick={() => handleUnassignCampaign(campaign.id)}
-                                    className="text-red-600 border-red-200 hover:bg-red-50 w-full sm:w-auto"
-                                  >
-                                    Remove
-                                  </Button>
-                                </div>
-                              );
-                            })}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-
-                    {/* Available Campaigns Section */}
-                    <div id="available-campaigns">
-                      <div className="flex items-center gap-2 mb-4">
-                        <Plus className="w-5 h-5 text-gray-400" />
-                        <h3 className="text-base sm:text-lg font-medium text-gray-900">Available Campaigns</h3>
-                        <Badge variant="secondary" className="bg-gray-100 text-gray-600 text-xs">
-                          {campaigns.filter(c => !newKiosk.assignedCampaigns.includes(c.id)).length}
-                        </Badge>
-                      </div>
-                      
-                      <div className="max-h-64 sm:max-h-80 overflow-y-auto">
-                        <div className="space-y-3">
-                          {campaigns
-                            .filter(campaign => !newKiosk.assignedCampaigns.includes(campaign.id))
-                            .map(campaign => {
-                              const fundingPercentage = Math.round((campaign.raised / campaign.goal) * 100);
-                              
-                              return (
-                                <div key={campaign.id} className="flex flex-col sm:flex-row sm:items-center gap-3 sm:gap-4 p-4 bg-gray-50 border border-gray-200 rounded-lg hover:bg-gray-100 transition-colors">
-                                  <div className="flex items-center gap-3 sm:gap-4 flex-1 min-w-0">
-                                    <div className="w-12 h-12 bg-white rounded-lg flex items-center justify-center flex-shrink-0 border">
-                                      {campaign.coverImageUrl ? (
-                                        <img 
-                                          src={campaign.coverImageUrl} 
-                                          alt={campaign.title}
-                                          className="w-full h-full object-cover rounded-lg"
-                                        />
-                                      ) : (
-                                        <span className="text-xs font-medium text-gray-500">
-                                          {campaign.title.substring(0, 3).toUpperCase()}
-                                        </span>
-                                      )}
-                                    </div>
-                                    
-                                    <div className="flex-1 min-w-0">
-                                      <h4 className="font-medium text-gray-900 truncate text-sm sm:text-base">{campaign.title}</h4>
-                                      <div className="flex flex-col sm:flex-row sm:items-center gap-1 sm:gap-4 text-xs sm:text-sm text-gray-500 mt-1">
-                                        <span>{formatCurrency(campaign.raised)} raised</span>
-                                        <span className="hidden sm:inline">•</span>
-                                        <span>{fundingPercentage}% funded</span>
-                                      </div>
-                                    </div>
-                                  </div>
-                                  
-                                  <Button
-                                    variant="outline"
-                                    size="sm"
-                                    onClick={() => handleAssignCampaign(campaign.id)}
-                                    className="text-green-600 border-green-200 hover:bg-green-50 w-full sm:w-auto"
-                                  >
-                                    <Plus className="w-4 h-4 mr-1" />
-                                    ASSIGN
-                                  </Button>
-                                </div>
-                              );
-                            })}
-                          
-                          {campaigns.filter(c => !newKiosk.assignedCampaigns.includes(c.id)).length === 0 && (
-                            <div className="text-center py-6 sm:py-8 text-gray-500">
-                              <p className="text-sm sm:text-base">All available campaigns have been assigned to this kiosk.</p>
-                            </div>
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </section>
-
-                {/* Display Section */}
-                <section 
-                  id="display"
-                  ref={sectionRefs['display']}
-                  className="p-4 sm:p-6 lg:p-8"
-                >
-                  <div className="max-w-4xl">
-                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-900 mb-6 sm:mb-8">
-                      Campaign View
-                    </h2>
-                    
-                    {/* Display Layout Selection */}
-                    <div className="mb-6 sm:mb-8">
-                      <h3 className="text-base sm:text-lg font-medium text-gray-900 mb-4">DISPLAY LAYOUT</h3>
-                      
-                      <div className="grid grid-cols-1 sm:grid-cols-3 gap-4 sm:gap-6 lg:gap-8">
-                        {/* Grid Option */}
-                        <button
-                          onClick={() => setNewKiosk(prev => ({ ...prev, displayLayout: 'grid' }))}
-                          className={`flex flex-col items-center justify-center w-full h-32 sm:h-36 rounded-2xl transition-all ${
-                            newKiosk.displayLayout === 'grid'
-                              ? 'border-2 border-green-500 bg-green-50'
-                              : 'border-2 border-gray-200 bg-transparent hover:bg-gray-50 hover:border-gray-300'
-                          }`}
-                          aria-pressed={newKiosk.displayLayout === 'grid'}
-                        >
-                          <Grid3X3 className={`w-8 h-8 sm:w-10 sm:h-10 mb-3 sm:mb-4 ${
-                            newKiosk.displayLayout === 'grid' ? 'text-green-600' : 'text-gray-400'
-                          }`} />
-                          <span className={`text-sm sm:text-base font-medium ${
-                            newKiosk.displayLayout === 'grid' ? 'text-green-600' : 'text-gray-500'
-                          }`}>
-                            GRID
-                          </span>
-                        </button>
-
-                        {/* List Option */}
-                        <button
-                          onClick={() => setNewKiosk(prev => ({ ...prev, displayLayout: 'list' }))}
-                          className={`flex flex-col items-center justify-center w-full h-32 sm:h-36 rounded-2xl transition-all ${
-                            newKiosk.displayLayout === 'list'
-                              ? 'border-2 border-green-500 bg-green-50'
-                              : 'border-2 border-gray-200 bg-transparent hover:bg-gray-50 hover:border-gray-300'
-                          }`}
-                          aria-pressed={newKiosk.displayLayout === 'list'}
-                        >
-                          <List className={`w-8 h-8 sm:w-10 sm:h-10 mb-3 sm:mb-4 ${
-                            newKiosk.displayLayout === 'list' ? 'text-green-600' : 'text-gray-400'
-                          }`} />
-                          <span className={`text-sm sm:text-base font-medium ${
-                            newKiosk.displayLayout === 'list' ? 'text-green-600' : 'text-gray-500'
-                          }`}>
-                            LIST
-                          </span>
-                        </button>
-
-                        {/* Carousel Option */}
-                        <button
-                          onClick={() => setNewKiosk(prev => ({ ...prev, displayLayout: 'carousel' }))}
-                          className={`flex flex-col items-center justify-center w-full h-32 sm:h-36 rounded-2xl transition-all ${
-                            newKiosk.displayLayout === 'carousel'
-                              ? 'border-2 border-green-500 bg-green-50'
-                              : 'border-2 border-gray-200 bg-transparent hover:bg-gray-50 hover:border-gray-300'
-                          }`}
-                          aria-pressed={newKiosk.displayLayout === 'carousel'}
-                        >
-                          <GalleryThumbnails className={`w-8 h-8 sm:w-10 sm:h-10 mb-3 sm:mb-4 ${
-                            newKiosk.displayLayout === 'carousel' ? 'text-green-600' : 'text-gray-400'
-                          }`} />
-                          <span className={`text-sm sm:text-base font-medium ${
-                            newKiosk.displayLayout === 'carousel' ? 'text-green-600' : 'text-gray-500'
-                          }`}>
-                            CAROUSEL
-                          </span>
-                        </button>
-                      </div>
-                    </div>
-
-                    {/* Layout Preview */}
-                    <div className="mb-6 sm:mb-8">
-                      <h3 className="text-base sm:text-lg font-medium text-gray-900 mb-4">LAYOUT PREVIEW</h3>
-                      
-                      <div className="bg-white border border-gray-200 rounded-xl p-4 sm:p-6 min-h-[300px] sm:min-h-[400px]">
-                        {/* Mock Device Frame */}
-                        <div className="bg-gray-100 rounded-lg p-3 sm:p-4 max-w-2xl mx-auto">
-                          {/* Device Header */}
-                          <div className="flex items-center justify-between mb-3 sm:mb-4 text-xs text-gray-500">
-                            <span>9:41 AM</span>
-                            <div className="flex gap-1">
-                              <div className="w-1 h-1 bg-gray-400 rounded-full"></div>
-                              <div className="w-1 h-1 bg-gray-400 rounded-full"></div>
-                              <div className="w-1 h-1 bg-gray-400 rounded-full"></div>
-                            </div>
-                          </div>
-                          
-                          {/* Content Area */}
-                          <div className="bg-white rounded-lg p-3 sm:p-4 min-h-[200px] sm:min-h-[300px]">
-                            {newKiosk.displayLayout === 'grid' && (
-                              <div className="grid grid-cols-2 gap-3 sm:gap-4">
-                                {[1, 2, 3, 4].map((item) => (
-                                  <div key={item} className="aspect-square bg-gray-100 rounded-lg flex items-center justify-center">
-                                    <Image className="w-6 h-6 sm:w-8 sm:h-8 text-gray-400" />
-                                  </div>
-                                ))}
-                              </div>
-                            )}
-                            
-                            {newKiosk.displayLayout === 'list' && (
-                              <div className="space-y-3 sm:space-y-4">
-                                {[1, 2, 3, 4].map((item) => (
-                                  <div key={item} className="flex items-center gap-3 sm:gap-4 p-2 sm:p-3 bg-gray-50 rounded-lg">
-                                    <div className="w-10 h-10 sm:w-12 sm:h-12 bg-gray-200 rounded-lg flex items-center justify-center flex-shrink-0">
-                                      <Image className="w-4 h-4 sm:w-6 sm:h-6 text-gray-400" />
-                                    </div>
-                                    <div className="flex-1">
-                                      <div className="h-2 sm:h-3 bg-gray-200 rounded mb-1 sm:mb-2"></div>
-                                      <div className="h-1.5 sm:h-2 bg-gray-200 rounded w-2/3"></div>
-                                    </div>
-                                    <div className="w-12 h-6 sm:w-16 sm:h-8 bg-green-200 rounded"></div>
-                                  </div>
-                                ))}
-                              </div>
-                            )}
-                            
-                            {newKiosk.displayLayout === 'carousel' && (
-                              <div className="relative">
-                                <div className="aspect-video bg-gray-100 rounded-lg flex items-center justify-center mb-3 sm:mb-4">
-                                  <Image className="w-12 h-12 sm:w-16 sm:h-16 text-gray-400" />
-                                </div>
-                                <div className="flex justify-center gap-2 mb-3 sm:mb-4">
-                                  <div className="w-2 h-2 bg-green-500 rounded-full"></div>
-                                  <div className="w-2 h-2 bg-gray-300 rounded-full"></div>
-                                  <div className="w-2 h-2 bg-gray-300 rounded-full"></div>
-                                </div>
-                                <div className="text-center">
-                                  <div className="h-3 sm:h-4 bg-gray-200 rounded mb-2 mx-auto w-3/4"></div>
-                                  <div className="h-2 sm:h-3 bg-gray-200 rounded mx-auto w-1/2"></div>
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                          
-                          {/* Bottom Navigation */}
-                          <div className="flex justify-center mt-3 sm:mt-4 gap-2">
-                            <div className="w-10 h-1.5 sm:w-12 sm:h-2 bg-gray-800 rounded-full"></div>
-                            <div className="w-6 h-1.5 sm:w-8 sm:h-2 bg-green-400 rounded-full"></div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </section>
-              </div>
-
-              {/* Footer - Fixed */}
-              <footer className="flex flex-col sm:flex-row items-center justify-between p-4 sm:p-6 lg:p-8 border-t border-gray-200 bg-gray-50 gap-4 sm:gap-0 flex-shrink-0">
-                <Button
-                  variant="ghost"
-                  onClick={() => {
-                    setIsCreateDialogOpen(false);
-                    setActiveSection('basic-info');
-                    setIsMobileSidebarOpen(false);
-                  }}
-                  className="text-gray-600 hover:text-gray-800 w-full sm:w-auto order-2 sm:order-1"
-                >
-                  <Save className="w-4 h-4 mr-2" />
-                  SAVE DRAFT
-                </Button>
-                
-                <div className="flex flex-col sm:flex-row gap-3 w-full sm:w-auto order-1 sm:order-2">
-                  <Button
-                    variant="outline"
-                    onClick={() => {
-                      setIsCreateDialogOpen(false);
-                      setActiveSection('basic-info');
-                      setIsMobileSidebarOpen(false);
-                    }}
-                    className="px-6 w-full sm:w-auto"
-                  >
-                    CANCEL
-                  </Button>
-                  <Button
-                    onClick={handleCreateKiosk}
-                    disabled={!newKiosk.name || !newKiosk.location}
-                    className="bg-black hover:bg-gray-800 text-white px-6 w-full sm:w-auto h-12 sm:h-auto"
-                  >
-                    <Save className="w-4 h-4 mr-2" />
-                    {editingKiosk ? 'UPDATE KIOSK' : 'SAVE KIOSK'}
-                  </Button>
-                </div>
-              </footer>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
+        editingKiosk={editingKiosk}
+        kioskData={newKiosk}
+        setKioskData={setNewKiosk}
+        campaigns={campaigns}
+        onSubmit={handleCreateKiosk}
+        onCancel={handleCancel}
+        onAssignCampaign={handleAssignCampaign}
+        onUnassignCampaign={handleUnassignCampaign}
+        formatCurrency={formatCurrency}
+      />
       
       {/* Delete Confirmation Dialog */}
       <Dialog open={isDeleteDialogOpen} onOpenChange={setIsDeleteDialogOpen}>
